Add New Pattern Program HourglassPattern.py
This is Hourglass Pattern program this program is a typical structure to print Hourglass pattern, i.e., the number of rows and columns. We need to use many loops to print this pattern, i.e., use nested loops. this program will help people to understand nesting of loops and some more patterns